Update getbambed.R
This sort of ref can change the order of ref compared to the bedFile, at least in the targeted pipeline. When the gene column of bedFile is cbind to ref, it will cause the genes to be mapped incorrectly. I see no reason to sort the ref; alternatively you can sort the bedFile the same way as ref each time you read it.